[Python] Install Python bindings with pip if available
As directly invoking `python setup.py` is considered deprecated behavior
by the PyPA, use `python -m pip install` to install the Python bindings
if the Python interpreter has a valid version of pip associated with it.

If there is a valid pip, update pip to the latest release and ensure that
setuptools and wheel are installed. Then build and install the Python bindings
with `python -m pip install`.

If there is no valid pip, warn the user and give them instructions on how to
get pip on their machine. Attempt to still install the Python bindings by falling
back on the deprecated `python setup.py build` and `python setup.py install` which
will also invoke Easy Install if setuptools v0.60.0+ is used.