[CI]: Add GitHub Actions based CI
Use GitHub Actions to add a CI workflow that runs on:

* push events to master or on tags
* pull request events targeting master
* a weekly CRON job that runs every Sunday at 01:23 UTC
* on GitHub releases
* on demand against any branch through workflow dispatch

Using concurrency groups if an event triggers another run of the workflow
while a previous run in ongoing, the older run will be automaticcally
canceled to avoid wasting resources.

The jobs targeti CentOS based workflow to start off with:

* Building with CMake like a user would and then running some libXrdCl tests.
This build is a shorter build.
* Building with CMake like a user would but with Python 3 pip, setuptools, and
wheel updated to the latest releases to check that users running with modern
releases won't hit breaking changes.
* Building in full RPM.
* Building a Python sdist with the publishing workflow scripts.
* Building a Python sdist with the publishing workflow scripts, but with
Python 3 pip, setuptools, and wheel updated to the latest releases to check
that users running with modern releases won't hit breaking changes.