[Python] Use 'wheel' to convert egg distribution to wheel
Instead of manually trying to convert an egg distribution to a wheel, using
`wheel`'s 'convert' CLI API to convert the egg into a wheel.

c.f. https://wheel.readthedocs.io/en/stable/reference/wheel_convert.html

As the egg distribution is formed during the installation, uninstall the egg
and then install the wheel in its place. This procedure in 'packaging/wheel/install.sh'
introduces a dependency on the the POSIX command 'find', and also introduces a
dependency on pip.